dev: split mock test harness into crate-local modules
